Refactored to make it easier to update
